third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-custom-expected.txt


----- tests for getting/setting interesting properties -----

PASS: canGet('targetWindow.Object') should be 'false' and is.
PASS: canSet('targetWindow.Object') should be 'false' and is.
PASS: canGet('targetWindow.existingCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.existingCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.newCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.hasOwnProperty') should be 'false' and is.
PASS: canSet('targetWindow.hasOwnProperty') should be 'false' and is.
PASS: canGet('targetWindow.windowPrototypeCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.windowPrototypeCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.objectPrototypeCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.objectPrototypeCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.toString') should be 'false' and is.
PASS: canSet('targetWindow.toString') should be 'false' and is.
PASS: toString('targetWindow') should be '[exception]' and is.
